requirements.txt
argon2-cffi==20.1.0
bcrypt==3.1.7
django==2.2.13
django-cors-headers==3.2.1
django-phonenumber-field
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pillow==7.1.2
pyjwt==1.7.1
